requirements.txt
# Binance & trading
python-binance==1.0.19
ccxt==4.2.25

# Async HTTP
aiohttp==3.9.5

# PostgreSQL (async)
asyncpg==0.29.0

# Veri işleme
pandas==2.2.2
numpy==1.26.4
ta==0.11.0

# Haber
requests==2.31.0
feedparser==6.0.11

# Güvenli ortam değişkenleri
python-dotenv==1.0.1

# Telegram bildirimleri (isteğe bağlı)
python-telegram-bot==21.3

# Loglama
rich==13.7.1
